Added the Region Check
Now checks for left hand’s current position using global “region”.
Costs are assigned accordingly.
Results of both Joker and the Thief and Stairway to Heaven are added to
output, seed value of first note needs to be changed every time and
tab[] needs to be reset to i for i in tuning.